requirements.txt
fastapi
uvicorn
python-docx
openai
python-dotenv
zeep
